Split NesteggPacketHolder from WebMReader
So both WebMReader and WebMDemuxer can use it without duplicating code.
